chore(appium,types): move DriverType, PluginType etc. into @appium/types
- Also split out types into more files, since they were getting hefty
- The `Plugin` type conflicts with the [Plugin API](https://developer.mozilla.org/en-US/docs/Web/API/Plugin), which is detailed in TypeScript's DOM library.  This changes makes `@appium/types` _avoid_ this library entirely, and only use the basic JS/Node.js types.

# Conflicts:
#	packages/appium/types/extension.ts
#	packages/types/lib/index.ts
