+++ packages/appium/lib/cli/extension-command.js
@@ -641,9 +641,9 @@ export {ExtensionCommand};
  */
 
 /**
- * @typedef {import('appium/types').ExtensionType} ExtensionType
- * @typedef {import('appium/types').DriverType} DriverType
- * @typedef {import('appium/types').PluginType} PluginType
+ * @typedef {import('@appium/types').ExtensionType} ExtensionType
+ * @typedef {import('@appium/types').DriverType} DriverType
+ * @typedef {import('@appium/types').PluginType} PluginType
  */
 
 /**
