+++ packages/appium/test/e2e/cli.e2e.spec.js
@@ -524,10 +524,10 @@ describe('CLI behavior', function () {
 });
 
 /**
- * @typedef {import('appium/types').ExtensionType} ExtensionType
+ * @typedef {import('@appium/types').ExtensionType} ExtensionType
  * @typedef {import('appium/types').ManifestData} ManifestData
- * @typedef {import('appium/types').DriverType} DriverType
- * @typedef {import('appium/types').PluginType} PluginType
+ * @typedef {import('@appium/types').DriverType} DriverType
+ * @typedef {import('@appium/types').PluginType} PluginType
  * @typedef {import('appium/lib/cli/extension-command').ExtensionListData} ExtensionListData
  * @typedef {import('./e2e-helpers').CliArgs} CliArgs
  * @typedef {import('appium/types').CliExtensionSubcommand} CliExtensionSubcommand
