+++ packages/appium/test/unit/extension/driver-config.spec.js
@@ -344,16 +344,16 @@ describe('DriverConfig', function () {
 });
 
 /**
- * @typedef {import('appium/types').DriverType} DriverType
+ * @typedef {import('@appium/types').DriverType} DriverType
  * @typedef {import('appium/lib/extension/driver-config').DriverConfig} DriverConfig
  */
 
 /**
- * @template {import('appium/types').ExtensionType} ExtType
+ * @template {import('@appium/types').ExtensionType} ExtType
  * @typedef {import('appium/types').ExtManifestWithSchema<ExtType>} ExtManifestWithSchema
  */
 
 /**
- * @template {import('appium/types').ExtensionType} ExtType
+ * @template {import('@appium/types').ExtensionType} ExtType
  * @typedef {import('appium/types').ExtManifest<ExtType>} ExtManifest
  */
