+++ packages/appium/types/index.ts
@@ -2,9 +2,6 @@ export * from './appium-manifest';
 export * from './external-manifest';
 export * from './extension';
 export * from './cli';
-export type DriverType = 'driver';
-export type PluginType = 'plugin';
-export type ExtensionType = DriverType | PluginType;
 
 /**
  * Known environment variables concerning Appium
