+++ packages/types/tsconfig.json
@@ -5,10 +5,9 @@
     "outDir": "build",
     "paths": {
       "@appium/schema": ["../schema"]
-    }
+    },
+    "lib": ["ES2015"]
   },
   "include": ["lib"],
-  "references": [
-    {"path": "../schema"}
-  ]
+  "references": [{"path": "../schema"}]
 }
